requirements.txt
aniso8601==8.0.0
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
cryptography==2.8
dnspython==1.16.0
Flask==0.12.4
Flask-RESTful==0.3.7
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.2
pip==19.0.3
pycparser==2.19
pytz==2019.3
setuptools==40.8.0
six==1.13.0
speaklater==1.3
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
python-telegram-bot
requests
redis